interim itinerary validation

+++ ap/interim/views.py
@@ -43,26 +43,34 @@ class InterimIntentionsView(UpdateView):
     start_list = data.pop('start')
     end_list = data.pop('end')
     commments_list = data.pop('comments')
-    InterimItinerary.objects.filter(interim_intentions=interim_intentions).delete()
+    itins = []
 
     for index in range(len(start_list)):
-      itin = InterimItinerary()
-      itin.interim_intentions = interim_intentions
-      itin.start = parser.parse(start_list[index])
-      itin.end = parser.parse(end_list[index])
-      itin.comments = commments_list[index]
-      itin.save()
+      itins.append(InterimItineraryForm(data={'start':start_list[index], 'end':end_list[index], 'comments':commments_list[index], 'interim_intentions': interim_intentions}))
+    if all(f.is_valid() for f in itins):
+      InterimItinerary.objects.filter(interim_intentions=interim_intentions).delete()
+      for itin in itins:
+        itobj = itin.save(commit=False)
+        itobj.interim_intentions = interim_intentions
+        itobj.save()
+    return itins
+
 
   def get_context_data(self, **kwargs):
     ctx = super(InterimIntentionsView, self).get_context_data(**kwargs)
     admin, created = InterimIntentionsAdmin.objects.get_or_create(term=Term.current_term())
     interim_itineraries_forms = []
-    interim_itineraries = InterimItinerary.objects.filter(interim_intentions=self.get_object()).order_by('start')
-    if interim_itineraries.count() == 0:
-      interim_itineraries_forms.append(InterimItineraryForm())
-    else:
-      for itin in interim_itineraries:
-        interim_itineraries_forms.append(InterimItineraryForm(instance=itin))
+    if self.request.method == 'POST':
+      interim_itineraries_forms = self.update_interim_itinerary(interim_intentions=self.get_object(), data=self.request.POST.copy())
+      print interim_itineraries_forms
+    elif self.request.method == 'GET':
+      interim_itineraries = InterimItinerary.objects.filter(interim_intentions=self.get_object()).order_by('start')
+      print interim_itineraries
+      if interim_itineraries.count() == 0:
+        interim_itineraries_forms.append(InterimItineraryForm())
+      else:
+        for itin in interim_itineraries:
+          interim_itineraries_forms.append(InterimItineraryForm(instance=itin))
     ctx['button_label'] = 'Submit'
     ctx['page_title'] = 'Interim Intentions'
     ctx['itinerary_forms'] = interim_itineraries_forms
